% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_site_github_with_extra_files.R
\name{construct_commit_message}
\alias{construct_commit_message}
\title{construct_commit_message}
\usage{
construct_commit_message(pkg, commit = Sys.getenv("TRAVIS_COMMIT"))
}
\description{
construct_commit_message
}
\keyword{internal}
