% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_site_github_with_extra_files.R
\name{copy_files_from_vignettes_dir_to_deploy_dir}
\alias{copy_files_from_vignettes_dir_to_deploy_dir}
\title{Copy files from Vignettes Dir to Deploy idir}
\usage{
copy_files_from_vignettes_dir_to_deploy_dir(source_dir = ".",
  deploy_dir = "docs", pattern = "\\\\.json$", overwrite = TRUE)
}
\arguments{
\item{source_dir}{default: "."}

\item{deploy_dir}{default: "docs}

\item{pattern}{file pattern to export (default: "\\.json$")}

\item{overwrite}{should existing files be overwritten (default: TRUE)}
}
\value{
files matching pattern copied to deploy_dir
}
\description{
Copy files from Vignettes Dir to Deploy idir
}
