% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_empty_branch.R
\name{create_empty_branch_ghpages}
\alias{create_empty_branch_ghpages}
\title{Create Empty gh-pages branch}
\usage{
create_empty_branch_ghpages(repo = NULL, org = "KWB-R",
  set_githubuser = TRUE, git_exe = path_to_git(),
  dest_dir = tempdir(), execute = TRUE, dbg = TRUE, ...)
}
\arguments{
\item{repo}{name of existing Github repository (default: NULL)}

\item{org}{repo owner (default: "KWB-R")}

\item{set_githubuser}{should the Github user be set before executing the
branch creation (default: TRUE). In this case set_github_user() is run}

\item{git_exe}{path to GIT executable (default:
kwb.pkgbuild:::path_to_git()), only required on "windows". On all other
platforms it is assumed that "git" is sufficient!}

\item{dest_dir}{directory where the the repo should be checked out (default:
tempdir())}

\item{execute}{should a batch file be run?}

\item{dbg}{print debug messages (default: TRUE)}

\item{...}{additional arguments passed to set_github_user}
}
\value{
create and push an empty gh-pages branch to org/repo
}
\description{
Create Empty gh-pages branch
}
