% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pkg_dir.R
\name{create_pkg_dir}
\alias{create_pkg_dir}
\title{Create Package Directory}
\usage{
create_pkg_dir(pkg_dir)
}
\arguments{
\item{pkg_dir}{path to package directory to be created, including the
pkgname as last folder}
}
\value{
creates directory for package if not existing
}
\description{
Create Package Directory
}
\examples{
## not a valid pkg folder
\dontrun{
pkg_dir <- file.path(tempdir(), "pkgname/pkgname")
create_pkg_dir(pkg_dir)
}
## valid pkg folder
pkg_dir <- file.path(tempdir(), "pkgname")
create_pkg_dir(pkg_dir)
}
