% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_autopkgdown.R
\name{create_travis_autopkgdown_yml}
\alias{create_travis_autopkgdown_yml}
\title{Create Travis Auto Pkgdown Yml}
\usage{

  create_travis_autopkgdown_yml(travis_template = travis_autopkgdown_yml(),
  dbg = TRUE)
}
\arguments{
\item{travis_template}{default: travis_autopkgdown_yml()}

\item{dbg}{(default: TRUE)}
}
\value{
".travis.yml" with content specified in travis_autopkgdown_yml()
}
\description{
Create Travis Auto Pkgdown Yml
}
