% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_site_github_with_extra_files.R
\name{deploy_local_with_extra_files}
\alias{deploy_local_with_extra_files}
\title{deploy_local_with_extra_files}
\usage{
deploy_local_with_extra_files(pkg = ".",
  vignettes_file_pattern_to_copy = "\\\\.json$", repo_slug = NULL,
  commit_message = pkgdown:::construct_commit_message(pkg), ...)
}
\arguments{
\item{pkg}{"."}

\item{vignettes_file_pattern_to_copy}{file patern for copying files from
vignettes directory into deploy directory (default: "\\.json$")}

\item{repo_slug}{repo_slug Sys.getenv("TRAVIS_REPO_SLUG", "")}

\item{commit_message}{commit_message pkgdown:::construct_commit_message(pkg)}

\item{...}{additional arguments passed to  pkgdown::build_site()}
}
\value{
deploy pkgdown site including additional files in "vignettes" source
folder to "gh-pages"
}
\description{
for details see pkgdown::deploy_local(), only parameter
vignettes_file_pattern_to_copy added
}
