% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_site_github_with_extra_files.R
\name{deploy_site_github_with_extra_files}
\alias{deploy_site_github_with_extra_files}
\title{deploy_site_github_with_extra_files}
\usage{
deploy_site_github_with_extra_files(
  pkg = ".",
  vignettes_file_pattern_to_copy = "\\\\.json$",
  tarball = Sys.getenv("PKG_TARBALL", ""),
  ssh_id = Sys.getenv("id_rsa", ""),
  repo_slug = Sys.getenv("TRAVIS_REPO_SLUG", ""),
  commit_message = pkgdown:::construct_commit_message(pkg),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{pkg}{"."}

\item{vignettes_file_pattern_to_copy}{file patern for copying files from
vignettes directory into deploy directory (default: "\\.json$")}

\item{tarball}{tarball Sys.getenv("PKG_TARBALL", "")}

\item{ssh_id}{ssh_id Sys.getenv("id_rsa", "")}

\item{repo_slug}{repo_slug Sys.getenv("TRAVIS_REPO_SLUG", "")}

\item{commit_message}{commit_message pkgdown:::construct_commit_message(pkg)}

\item{verbose}{verbose (default: TRUE)}

\item{...}{additional arguments passed to deploy_local_with_extra_files()}
}
\value{
deploy pkgdown site including additional files in "vignettes" source
folder to "gh-pages"
}
\description{
for details see pkgdown::deploy_site_github(), only parameter
vignettes_file_pattern_to_copy added
}
