% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_site_github_with_extra_files.R
\name{deploy_to_branch_with_extra_files}
\alias{deploy_to_branch_with_extra_files}
\title{Build and deploy a site locally with extra files}
\usage{
deploy_to_branch_with_extra_files(
  pkg = ".",
  vignettes_file_pattern_to_copy = "\\\\.json",
  commit_message = construct_commit_message(pkg),
  clean = FALSE,
  branch = "gh-pages",
  remote = "origin",
  github_pages = (branch == "gh-pages"),
  ...
)
}
\arguments{
\item{pkg}{"."}

\item{vignettes_file_pattern_to_copy}{file patern for copying files from
vignettes directory into deploy directory (default: "\\.json$")}

\item{commit_message}{commit_message (default: construct_commit_message(pkg))}

\item{clean}{Clean all files from old site (default: FALSE)}

\item{branch}{The git branch to deploy to}

\item{remote}{The git remote to deploy to}

\item{github_pages}{Is this a GitHub pages deploy. If `TRUE`, adds a `CNAME`
file for custom domain name support, and a `.nojekyll` file to suppress
jekyll rendering.}

\item{...}{Additional arguments passed to pkgdown::build_site()}
}
\description{
Assumes that you're in a git clone of the project, and the package is
already installed.
}
