% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_config.R
\name{git_check_config}
\alias{git_check_config}
\title{Git Check Configuration}
\usage{
git_check_config()
}
\value{
prints 'global' and 'local' Git configuration
}
\description{
Git Check Configuration
}
\examples{
\dontrun{
git_check_config()
}
}
