% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_config.R
\name{git_setup_user}
\alias{git_setup_user}
\title{Git Setup User (Name and Email)}
\usage{
git_setup_user(github_username, github_email, scope = c("local",
  "global"), git_exe = path_to_git())
}
\arguments{
\item{github_username}{username should be your GitHub username}

\item{github_email}{email should be identical to email you registered with at
GitHub (e.g. your.name(at)kompetenz-wasser.de)}

\item{scope}{scope (default: c("local", "global"))}

\item{git_exe}{path to git executable on Windows (default: path_to_git())}
}
\value{
runs "git" to setup the Git(Hub) user
}
\description{
Git Setup User (Name and Email)
}
\examples{
\dontrun{
git_setup_user("mrustl", "michael.rustler@kompetenz-wasser.de")
}
}
\seealso{
<https://support.rstudio.com/hc/en-us/community/posts/115001143667-Author-Change-with-Git-in-RStudio>
}
