% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_description.R
\name{read_description}
\alias{read_description}
\title{Helper function: read_description}
\usage{
read_description(file = "DESCRIPTION")
}
\arguments{
\item{file}{path to DESCRIPTION file (default: DESCRIPTION)}
}
\value{
list with pkg "name", "title", "desc", "version"
}
\description{
Helper function: read_description
}
