% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{set_github_user}
\alias{set_github_user}
\title{Set Github User For GIT}
\usage{
set_github_user(git_username = "kwb.pkgbuild::use_autopkgdown()",
  git_fullname = "kwb.pkgbuild::use_autopkgdown()",
  git_email = "kwbr-bot@kompetenz-wasser.de",
  git_exe = "C:/Program Files/Git/bin/git.exe", execute = FALSE,
  execute_dir = tempdir(), dbg = TRUE)
}
\arguments{
\item{git_username}{Github username (default: "kwb.pkgbuild::use_autopkgdown()")}

\item{git_fullname}{Github fullname (default: "kwb.pkgbuild::use_autopkgdown()")}

\item{git_email}{(default: `kwbr-bot@kompetenz-wasser.de`), is then internally derived by calling
whoami::fullname() and assuming that Github "kompetenz-wasser.de" used as
email on Github}

\item{git_exe}{path git git executable (only used on Windows as it is not
installed in the local "environment"). On all other plattforms it is assumed
that a call with "git" works}

\item{execute}{should a batch file be run?}

\item{execute_dir}{path to directory where batch script should be run in case
that execute == TRUE (default: tempdir())}

\item{dbg}{print debug messages (default: TRUE)}
}
\value{
sets globally user.name and user.email in Git
}
\description{
Set Github User For GIT
}
