% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_autopkgdown.R
\name{travis_autopkgdown_yml}
\alias{travis_autopkgdown_yml}
\title{Use Travis Auto Pkgdown Yml Template}
\usage{
travis_autopkgdown_yml()
}
\value{
default yaml if Auto Pkgdown is used
}
\description{
Use Travis Auto Pkgdown Yml Template
}
\examples{
cat(travis_autopkgdown_yml(), sep = '\\n')

}
