% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_appveyor.R
\name{use_appveyor}
\alias{use_appveyor}
\title{Adds default appveyor.yml}
\usage{
use_appveyor(dest_dir = getwd(), yml_vector = ci_appveyor_template())
}
\arguments{
\item{dest_dir}{directoy to write (default: getwd())}

\item{yml_vector}{a yml imported as string vector (default:
ci_appveyor_template())}
}
\value{
writes appveyor.yml and adds it .Rbuildignore
}
\description{
Adds default appveyor.yml
}
