% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_autopkgdown.R
\name{use_autopkgdown}
\alias{use_autopkgdown}
\title{Use Auto Pkgdown}
\usage{
use_autopkgdown(repo = NULL, org = "KWB-R", set_githubuser = TRUE,
  git_exe = path_to_git(), dest_dir = tempdir(), execute = TRUE,
  dbg = TRUE, ...)
}
\arguments{
\item{repo}{name of existing Github repository (default: NULL)}

\item{org}{repo owner (default: "KWB-R")}

\item{set_githubuser}{should the Github user be set before executing the
branch creation (default: TRUE). In this case set_github_user() is run}

\item{git_exe}{path to GIT executable (default:
kwb.pkgbuild:::path_to_git()), only required on "windows". On all other
platforms it is assumed that "git" is sufficient!}

\item{dest_dir}{directory where the the repo should be checked out (default:
tempdir())}

\item{execute}{should a batch file be run?}

\item{dbg}{print debug messages (default: TRUE)}

\item{...}{additional arguments passed to set_github_user}
}
\value{
workflow required for automating pkgdown::build_site() on Travis
}
\description{
automate pkgdown::build_site by running it on Travis and
deploying to gh-pages branch
}
