% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_badges.R
\name{use_badge_appveyor}
\alias{use_badge_appveyor}
\title{Badge appveyor}
\usage{
use_badge_appveyor(repo = NULL, user = "KWB-R", domain = "github")
}
\arguments{
\item{repo}{name of repository (default: NULL)}

\item{user}{user name or organisation under which repository defined in
parameter "repo" is hosted (default: "KWB-R")}

\item{domain}{under which repository is hosted (default: "github")}
}
\value{
generates appveyor badge link
}
\description{
Badge appveyor
}
