% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_badges.R
\name{use_badge_cran}
\alias{use_badge_cran}
\title{Badge CRAN}
\usage{
use_badge_cran(pkgname)
}
\arguments{
\item{pkgname}{name of R package}
}
\value{
generates CRAN badge link
}
\description{
Badge CRAN
}
