% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_badges.R
\name{use_badge_lifecycle}
\alias{use_badge_lifecycle}
\title{Badge lifecycle}
\usage{
use_badge_lifecycle(stage = "experimental")
}
\arguments{
\item{stage}{badge declares the developmental stage of a package, according
to
[https://www.tidyverse.org/lifecycle/](https://www.tidyverse.org/lifecycle/),
 valid arguments are: "experimental", "maturing", "stable", "retired",
"archived", "dormant", "questioning"), (default: "experiment")}
}
\value{
generates lifecycle badge link
}
\description{
Badge lifecycle
}
