% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_badges.R
\name{use_badge_travis}
\alias{use_badge_travis}
\title{Badge travis}
\usage{
use_badge_travis(repo = NULL, user = "KWB-R")
}
\arguments{
\item{repo}{name of repository (default: NULL)}

\item{user}{user name or organisation under which repository defined in
parameter "repo" is hosted (default: "KWB-R")}
}
\value{
generates travis badge link
}
\description{
Badge travis
}
