% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gitlab_ci.R
\name{use_gitlab_ci}
\alias{use_gitlab_ci}
\title{Adds .gitlab-ci.yml (if repo contains a "docs" subfolder)}
\usage{
use_gitlab_ci()
}
\value{
writes .gitlab-ci.yml and adds it .Rbuildignore
}
\description{
Adds .gitlab-ci.yml (if repo contains a "docs" subfolder)
}
