% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gitlab_ci.R
\name{use_gitlab_ci_docs}
\alias{use_gitlab_ci_docs}
\title{Adds .gitlab-ci.yml (if repo contains a "docs" subfolder)}
\usage{
use_gitlab_ci_docs(dest_dir = getwd(), yml_vector = gitlab_ci_template_docs())
}
\arguments{
\item{dest_dir}{directory to write (default: getwd())}

\item{yml_vector}{a yml imported as string vector (default:
gitlab_ci_template_docs())}
}
\value{
writes .gitlab-ci.yml and adds it .Rbuildignore
}
\description{
Adds .gitlab-ci.yml (if repo contains a "docs" subfolder)
}
