% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gitlab_ci.R
\name{use_gitlab_ci_ghpages}
\alias{use_gitlab_ci_ghpages}
\title{Adds .gitlab-ci.yml (which should be saved in root dir of "gh-pages" branch)}
\usage{
use_gitlab_ci_ghpages(
  dest_dir = getwd(),
  yml_vector = gitlab_ci_template_ghpages()
)
}
\arguments{
\item{dest_dir}{directory to write (default:
getwd())}

\item{yml_vector}{a yml imported as string vector (default: gitlab_ci_template_ghpages())}
}
\value{
writes .gitlab-ci.yml
}
\description{
Adds .gitlab-ci.yml (which should be saved in root dir of "gh-pages" branch)
}
