% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_index_rmd.R
\name{use_index_rmd}
\alias{use_index_rmd}
\title{Use index.Rmd (used for pkgdown::build_home())}
\usage{
use_index_rmd(user = "KWB-R", domain = "github",
  stage = "experimental")
}
\arguments{
\item{user}{user name or organisation under which repository defined in\cr
parameter "repo" is hosted (default: "KWB-R")\cr}

\item{domain}{under which repository is hosted (default: "github")}

\item{stage}{badge declares the developmental stage of a package, according
to [https://www.tidyverse.org/lifecycle/](https://www.tidyverse.org/lifecycle/),
valid arguments are: "experimental", "maturing", "stable", "retired",
"archived", "dormant", "questioning"), (default: "experiment")}
}
\value{
generates travis badge link
}
\description{
Use index.Rmd (used for pkgdown::build_home())
}
