% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_installation.R
\name{use_installation}
\alias{use_installation}
\title{Helper function: describe pkg installation in index.Rmd / README.md}
\usage{
use_installation(pkgname, user, domain)
}
\arguments{
\item{pkgname}{package name}

\item{user}{user name or organisation under which repository defined in
parameter "repo" is hosted}

\item{domain}{under which repository is hosted}
}
\description{
Helper function: describe pkg installation in index.Rmd / README.md
}
