% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_mit_license.R
\name{use_mit_license}
\alias{use_mit_license}
\title{Create MIT licence with KWB style}
\usage{
use_mit_license(
  copyright_holder = list(name = kwb_string(), start_year = NULL)
)
}
\arguments{
\item{copyright_holder}{list with name of copyright holder and additional
start_year (e.g. "2017") of copyright protection. If not specified
(start_year = NULL), the current year is used as year only!
(default: "list(name = kwb.pkgbuild:::kwb_string(), start_year = NULL)")}
}
\value{
creates MIT licence file
}
\description{
Create MIT licence with KWB style
}
