% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_news_md.R
\name{use_news_md}
\alias{use_news_md}
\title{Create KWB-styled "NEWS.md"}
\usage{

  use_news_md(news_txt = "* Added a `NEWS.md` file to track changes to the package.",
  style_guide_url = "https://style.tidyverse.org/news.html")
}
\arguments{
\item{news_txt}{text added to news (default: "* Added a 'NEWS.md' file to
track changes to the package.")}

\item{style_guide_url}{refer to tidyverse style website documenting how to
write a good "NEWS.md"(default: "http://style.tidyverse.org/news.html")}
}
\value{
writes "NEWS.md" (in case it is not existing)
}
\description{
Create KWB-styled "NEWS.md"
}
