% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pkg_skeleton.R
\name{use_pkg_skeleton}
\alias{use_pkg_skeleton}
\title{Use Package Skeleton}
\usage{
use_pkg_skeleton(pkg_name)
}
\arguments{
\item{pkg_name}{name of R package}
}
\value{
creates pkg skeleton in current working directory
}
\description{
Use Package Skeleton
}
\examples{
## valid pkg folder
pkg_name <- "pkgname"
pkg_dir <- file.path(tempdir(), pkg_name)
pkg_dir <- create_pkg_dir(pkg_dir)
withr::with_dir(pkg_dir, {use_pkg_skeleton(pkg_name)})
}
