% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pkgdown.R
\name{use_pkgdown}
\alias{use_pkgdown}
\title{Set up \code{pkgdown} with KWB styling}
\usage{
use_pkgdown(
  author = kwb_author("rustler"),
  copyright_holder_name = kwb_string(),
  pkg = get_pkgname(),
  user = "kwb-r",
  domain = "github"
)
}
\arguments{
\item{author}{list of author attributes (default:
kwb.pkgbuild:::kwb_author("rustler"))}

\item{copyright_holder_name}{name of copyright holder
(default: kwb.pkgbuild:::kwb_string())}

\item{pkg}{name of KWB package (default: get_pkgname())}

\item{user}{name of GitHub user/organisation (default: 'kwb-r')}

\item{domain}{name of domain for webpage publishing (default: 'github')}
}
\value{
invisibly; as a side effect writes \code{_pkgdown.yml} with KWB styling
and adds it to \code{.Rbuildignore}.
}
\description{
Calls \code{usethis::use_pkgdown()} and additionally writes a \code{_pkgdown.yml}
configured with KWB defaults (Bootstrap 5 + cerulean theme, KWB authors
block, copyright holder logo).
}
