% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_readme.R
\name{use_readme}
\alias{use_readme}
\title{Use README}
\usage{
use_readme(pkg = list(name = "kwb.lca", desc =
  paste("Helper functions for data aggregation and",
  "visualisation of UMBERTO (https://www.ifu.com/umberto)\\nmodel output")),
  user = "KWB-R", domain = "github", stage = "experimental")
}
\arguments{
\item{pkg}{package description in list format (default:\cr
pkg = list(name = "kwb.lca",\cr
title = "R package supporting UMERTO LCA at KWB",\cr
desc = "Helper functions for data aggregation and visualisation of\cr
UMBERTO (https://www.ifu.com/umberto) model output"))}

\item{user}{user name or organisation under which repository defined in
parameter "repo" is hosted (default: "KWB-R")}

\item{domain}{under which repository is hosted (default: "github")}

\item{stage}{badge declares the developmental stage of a package, according
to [https://www.tidyverse.org/lifecycle/](https://www.tidyverse.org/lifecycle/),
valid arguments are: "experimental", "maturing", "stable", "retired",
"archived", "dormant", "questioning"), (default: "experiment")}
}
\value{
generates travis badge link
}
\description{
Use README
}
