% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_readme_md.R
\name{use_readme_md}
\alias{use_readme_md}
\title{Create KWB-styled \code{README.md}}
\usage{
use_readme_md(user = "KWB-R", domain = "github", stage = "experimental")
}
\arguments{
\item{user}{user name or organisation under which the repository is hosted
(default: "KWB-R")}

\item{domain}{under which the repository is hosted (default: "github")}

\item{stage}{badge declaring the developmental stage of the package
according to
\url{https://www.tidyverse.org/lifecycle/};
valid values are "experimental", "maturing", "stable", "retired",
"archived", "dormant", "questioning" (default: "experimental")}
}
\value{
writes \code{README.md} and adds it to \code{.Rbuildignore}
}
\description{
Generates a \code{README.md} with the KWB default badge set, the package
description from \code{DESCRIPTION}, an installation snippet and links to the
release and development documentation websites.
}
