% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_travis.R
\name{use_travis}
\alias{use_travis}
\title{Adds default .travis.yml}
\usage{
use_travis(cache_pkg = TRUE)
}
\arguments{
\item{cache_pkg}{should packages be cached? TRUE/FALSE (default: TRUE)}
}
\value{
writes .travis.yml and adds it .Rbuildignore
}
\description{
Adds default .travis.yml
}
