% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{write_to_rbuildignore}
\alias{write_to_rbuildignore}
\title{Write to .Rbuildignore}
\usage{
write_to_rbuildignore(ignore_pattern)
}
\arguments{
\item{ignore_pattern}{pattern that should be added to .Rbuildignore}
}
\value{
.Rbuildignore created/updated with pattern
}
\description{
Write to .Rbuildignore
}
