% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_pkgname}
\alias{get_pkgname}
\title{Helper Function: Get Package Name}
\usage{
get_pkgname(pkgname = NULL)
}
\arguments{
\item{pkgname}{either package name or NULL. In this
case the DESCRIPTION file in the current working
directory is read and is package name ues (default: NULL)}
}
\value{
package name
}
\description{
Helper Function: Get Package Name
}
