% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_badges.R
\name{use_badge_codecov}
\alias{use_badge_codecov}
\title{Badge codecov}
\usage{
use_badge_codecov(repo, user = "KWB-R", domain = "github")
}
\arguments{
\item{repo}{name of repository}

\item{user}{user name or organisation under which repository defined in
parameter "repo" is hosted (default: KWB-R")}

\item{domain}{under which repository is hosted (default: "github")}
}
\value{
generates codecov badge link
}
\description{
Badge codecov
}
