% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_badges.R
\name{use_badge_ghactions_rcmdcheck}
\alias{use_badge_ghactions_rcmdcheck}
\title{Badge Github Actions RCMD Check}
\usage{
use_badge_ghactions_rcmdcheck(repo = NULL, user = "KWB-R", branch = NULL)
}
\arguments{
\item{repo}{name of repository (default: NULL)}

\item{user}{user name or organisation under which repository defined in
parameter "repo" is hosted (default: KWB-R")}

\item{branch}{default: NULL (i.e. use "default" branch) or user defined branch
(e.g. "dev")}
}
\value{
generates travis badge link
}
\description{
Badge Github Actions RCMD Check
}
