% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_badges.R
\name{use_badge_runiverse}
\alias{use_badge_runiverse}
\title{Badge R-Universe}
\usage{
use_badge_runiverse(repo = NULL, user = "KWB-R")
}
\arguments{
\item{repo}{name of repository (default: NULL)}

\item{user}{user name or organisation under which repository defined in
parameter "repo" is hosted (default: KWB-R")}
}
\value{
generates codecov badge link
}
\description{
Badge R-Universe
}
