% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_description.R
\name{use_description}
\alias{use_description}
\title{Use DESCRIPTION with KWB styling}
\usage{
use_description(
  author = kwb_author("rustler"),
  pkg = kwb_package("kwb.umberto"),
  version = NULL,
  license = "MIT + file LICENSE",
  copyright_holder_name = kwb_string(),
  funder = NULL
)
}
\arguments{
\item{author}{author information in list format (default:
kwb.pkgbuild:::kwb_author("rustler"))}

\item{pkg}{package description in list format (default:
kwb.pkgbuild:::kwb_package("kwb.umberto"))}

\item{version}{user defined version number (e.g. 0.1.0) or default version
number 0.0.0.9000 in case (version = NULL) of first release}

\item{license}{license (default: "MIT + file LICENSE")}

\item{copyright_holder_name}{name of copyright holder (default:
kwb.pkgbuild:::kwb_string())}

\item{funder}{funder/funding agency of R package (default: NULL), e.g.
project name (e.g. AQUANES)}
}
\value{
writes DESCRIPTION file using usethis::use_description() with KWB
  style
}
\description{
Use DESCRIPTION with KWB styling
}
