% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_github_ci.R
\name{use_ghactions}
\alias{use_ghactions}
\title{Adds default .github/workflows/}
\usage{
use_ghactions()
}
\value{
writes .github/workflows/ and adds it .Rbuildignore
}
\description{
Adds default .github/workflows/
}
