% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gitlab_ci.R
\name{use_gitlab_ci_blogdown}
\alias{use_gitlab_ci_blogdown}
\title{Adds .gitlab-ci.yml (if repo contains on root in a "gh-pages" branch)}
\usage{
use_gitlab_ci_blogdown(
  dest_dir = getwd(),
  yml_vector = gitlab_ci_template_blogdown()
)
}
\arguments{
\item{dest_dir}{directoy to write (default: getwd())}

\item{yml_vector}{a yml imported as string vector (default:
gitlab_ci_template_blogdown())}
}
\value{
writes .gitlab-ci.yml
}
\description{
Adds .gitlab-ci.yml (if repo contains on root in a "gh-pages" branch)
}
