% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gitlab_ci.R
\name{use_gitlab_ci_ghpages}
\alias{use_gitlab_ci_ghpages}
\title{Adds .gitlab-ci.yml (if repo contains on root in a "gh-pages" branch)}
\usage{
use_gitlab_ci_ghpages(dest_dir = getwd(),
  yml_list = gitlab_ci_template_ghpages())
}
\arguments{
\item{dest_dir}{directoy to write (default:
getwd())}

\item{yml}{a valid yml list (default: gitlab_ci_template_ghpages())}
}
\value{
writes .gitlab-ci.yml and adds it .Rbuildignore
}
\description{
Adds .gitlab-ci.yml (if repo contains on root in a "gh-pages" branch)
}
