% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pkgdown.R
\name{use_pkgdown}
\alias{use_pkgdown}
\title{pkgdown for KWB}
\usage{
use_pkgdown(author = kwb_author("rustler"),
  copyright_holder_name = kwb_string())
}
\arguments{
\item{author}{list of author attributes (default:
kwb.pkgbuild:::kwb_author("rustler"))}

\item{copyright_holder_name}{name of copyright holder
(default: kwb.pkgbuild:::kwb_string())}
}
\value{
performs usethis::use_pkgdown() and additionally writes _pkgdown.yml
based on KWB styling
}
\description{
pkgdown for KWB
}
