% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pkgdown.R
\name{use_pkgdown}
\alias{use_pkgdown}
\title{pkgdown for KWB}
\usage{
use_pkgdown(author = list(name = "Michael Rustler", url =
  "http://mrustl.de"),
  copyright_holder_name = "Kompetenzzentrum Wasser Berlin gGmbH (KWB)")
}
\arguments{
\item{author}{list of author attributes (default: list(name = "Michael Rustler",
url = "http://mrustl.de"))}

\item{copyright_holder_name}{name of copyright holder
(default: "Kompetenzzentrum Wasser Berlin gGmbH (KWB)")}
}
\value{
performs usethis::use_pkgdown() and additionally writes _pkgdown.yaml
based on KWB styling
}
\description{
pkgdown for KWB
}
