% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_travis.R
\name{use_travis}
\alias{use_travis}
\title{Adds default .travis.yml}
\usage{
use_travis(dbg = TRUE, yml_vector = ci_travis_template())
}
\arguments{
\item{dbg}{print debug messages (default: TRUE)}

\item{yml_vector}{a yml imported as string vector (default:
ci_travis_template())}
}
\value{
writes .travis.yml and adds it .Rbuildignore
}
\description{
Adds default .travis.yml
}
