% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{write_to_gitignore}
\alias{write_to_gitignore}
\title{Write to .gitignore}
\usage{
write_to_gitignore(ignore_pattern, comment)
}
\arguments{
\item{ignore_pattern}{pattern that should be added to .Rbuildignore}

\item{comment}{optional comment to write to file before the ignore pattern}
}
\value{
.gitignore created/updated with pattern
}
\description{
Write to .gitignore
}
